1_Multi-Class Prediction of Cirrhosis Outcomes/Streamlit_App/requirements.txt
matplotlib==3.9.0
numpy==2.0.0
pandas==2.2.2
plotly==5.22.0
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.14.0
seaborn==0.13.2
streamlit==1.36.0
streamlit_extras==0.4.3
streamlit-aggrid==1.0.5
streamlit_option_menu==0.3.13

